+++ CHANGELOG.md
@@ -70,6 +70,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - utils: Fix incorrect check on inability to set affinity from an interpreter
+- utils: Remove unnecessary requirement to configure the agent process as root
 
 ### Removed